Reviewer: Li XianJing Bug No.: NONE Description: 1. Hi all:
In current implement of main loop select, the timeout of select has not been
checked.

When select return, it will go to check the fd and ISSET firstly:
if((fd = ftk_source_get_fd(source)) >= 0 && FD_ISSET(fd, &thiz->fdset))

And then check the timeout:
if((source_wait_time = ftk_source_check(source)) == 0)

This will be a problem since the select is returned by timeout.

There is a sequence issue for timeout and fd checking in some situations:
If a timeout reached, but a fd is readable after select timed out in the
checking of FD_ISSET, it will go to fd section, and will not enter the timeout
section.
So the timeout ftk_source_dispatch will not be called.

So I make a patch for checking the "ret > 0" in fd section.

It needs a discussion before I commit this patch.
Thanks.

Tao Yu <yut616@gmail.com>